Разделены рисование поля (Field.kt) и данные поля (FieldData.kt)
